requirements.txt
librosa>=0.6.2
pytorch>=1.0
tqdm>=4.31.1
